Solved problem with 0ad game.
I figured out that game logic is implemented in JS programming language and
that previous integer location of was updated every time before usage.

JS represents all numbers as IEEE 64-bit double.
That is why searchig for 64-bit double instead of integer totally solved the
problem.

It seems that many games use C++ engine and some high-level scripting language.
That's why it is worth searching for floating-point numbers and probably
some unicode strings.

All the code is to be rewrited in a normal way.